fixed payment/guest relationship

+++ ITSC3155-Assignments-main 2/FinalProject/api/models/payments.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, ForeignKey, Integer, String, DECIMAL, DATETIME
+from sqlalchemy import Column, ForeignKey, Integer, String, BIGINT, DECIMAL, DATETIME
 from sqlalchemy.orm import relationship
 from datetime import datetime
 from ..dependencies.database import Base
@@ -8,7 +8,7 @@ class Payment(Base):
 
     id = Column(Integer, primary_key=True, index=True, autoincrement=True)
     guest_id = Column(Integer, ForeignKey("guests.id"))
-    card_number = Column(Integer, unique=True, nullable=False)
+    card_number = Column(BIGINT, unique=True, nullable=False)
     pin = Column(Integer, unique=True, nullable=True)
     method = Column(String(5), nullable=False)
